a ton of refactoring, no intended changes to features
