+++ README.md
@@ -1,0 +1,56 @@
+# MiniLoom
+
+MiniLoom is an AI text generation tool built with Electron, providing a desktop application for text generation and manipulation.
+
+## Features
+
+- Desktop application with native UI
+- Text generation capabilities
+- File save/load functionality
+- Settings management
+- Search functionality
+
+## Prerequisites
+
+- Node.js (version 14 or higher)
+- npm (comes with Node.js)
+
+## Installation
+
+1. Clone the repository:
+
+```bash
+git clone <repository-url>
+cd miniloom
+```
+
+2. Install dependencies:
+
+```bash
+npm install
+```
+
+## Usage
+
+To start the application:
+
+```bash
+npm start
+```
+
+This will launch the MiniLoom desktop application.
+
+## Development
+
+The application is built with:
+
+- Electron for the desktop framework
+- Vanilla JavaScript for the frontend
+- MiniSearch for search functionality
+- Diff-match-patch for text comparison
+
+## License
+
+Apache License 2.0
+
+This project is licensed under the Apache License 2.0 - see the [LICENSE](LICENSE) file for details.
