+++ src/index.html
@@ -76,18 +76,9 @@
         border-left: 1px solid #e9ecef;
         overflow-y: auto;
         padding: 16px;
-      }
-
-      /* Override injected styles from minisearch-integration.js */
-      #tree-container {
-        width: 100% !important;
-        margin-left: 0 !important;
-        margin-top: 0 !important;
-      }
-
-      #loom-tree-view {
-        margin-left: 0 !important;
-        width: 100% !important;
+        display: flex;
+        flex-direction: column;
+        height: 100vh;
       }
 
       #narrative-pane {
@@ -338,12 +329,6 @@
         flex-shrink: 0;
       }
 
-      .die-and-generate-container {
-        display: flex;
-        align-items: center;
-        gap: 8px;
-      }
-
       .generate-button-container .reroll {
         width: auto !important;
         max-width: none !important;
@@ -517,6 +502,7 @@
           -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif;
         font-size: 14px;
         line-height: 1.3;
+        min-height: 0;
       }
 
       #loom-tree-view ul {
@@ -665,13 +651,6 @@
         color: red;
       }
 
-      /* Search-specific styles */
-      #search-input {
-        text-align: left !important;
-        width: calc(100% - 16px) !important;
-        margin: 0 !important;
-      }
-
       .search-result-item {
         border: 1px solid #ddd;
         margin: 5px 0;
@@ -727,7 +706,79 @@
         border-bottom: none;
       }
 
+      .search-suggestion-highlighted {
+        background-color: #e0e0e0 !important;
+      }
+
       #search-suggestions {
+        box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+      }
+
+      /* Additional search result styles */
+      .search-no-results {
+        padding: 10px;
+        color: #666;
+        font-style: italic;
+        margin-top: 10px;
+        flex: 1;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        min-height: 0;
+      }
+
+      .search-results-container {
+        flex: 1;
+        overflow-y: auto;
+        margin-top: 10px;
+        min-height: 0;
+      }
+
+      .search-result-date {
+        margin-top: 2px;
+        color: #999;
+      }
+
+      /* Search container and input styles */
+      #search-container {
+        margin-bottom: 10px;
+        padding: 0;
+        position: relative;
+        width: 100%;
+        box-sizing: border-box;
+        flex-shrink: 0;
+      }
+
+      #search-input {
+        width: 100%;
+        padding: 8px;
+        border: 1px solid #ddd;
+        border-radius: 3px;
+        font-size: 14px;
+        box-sizing: border-box;
+        text-align: left !important;
+      }
+
+      #search-input:focus {
+        outline: none;
+        border-color: #007bff;
+        box-shadow: 0 0 0 2px rgba(0, 123, 255, 0.25);
+      }
+
+      #search-suggestions {
+        position: absolute;
+        top: 100%;
+        left: 0;
+        right: 0;
+        background: white;
+        border: 1px solid #ddd;
+        border-top: none;
+        border-radius: 0 0 3px 3px;
+        max-height: 150px;
+        overflow-y: auto;
+        z-index: 1000;
+        display: none;
+        box-sizing: border-box;
         box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
       }
 
@@ -789,12 +840,6 @@
       .control-group select:hover {
         border-color: #adb5bd;
       }
-
-      #die {
-        height: 32px;
-        width: 40px;
-        margin-right: 8px;
-      }
     </style>
   </head>
   <body>
@@ -804,7 +849,7 @@
           <div class="top-bar-content">
             <div class="logo-title">
               <img
-                src="assets/minihf_logo_no_text.png"
+                src="../assets/minihf_logo_no_text.png"
                 alt="MiniHF Logo"
                 class="app-logo"
               />
@@ -878,10 +923,28 @@
         </div>
       </div>
       <div id="nav-pane">
+        <div id="search-container">
+          <input type="text" id="search-input" placeholder="Search nodes..." />
+          <div id="search-suggestions"></div>
+        </div>
         <div id="loom-tree-view"></div>
+        <!-- Search result containers - hidden by default -->
+        <div
+          id="search-no-results"
+          class="search-no-results"
+          style="display: none"
+        >
+          No results found
+        </div>
+        <div
+          id="search-results-container"
+          class="search-results-container"
+          style="display: none"
+        ></div>
       </div>
     </div>
+    <script src="utils.js"></script>
     <script src="renderer.js"></script>
-    <script src="minisearch-integration.js"></script>
+    <script src="search.js"></script>
   </body>
 </html>
