+++ src/settings.html
@@ -654,6 +654,7 @@
     </div>
 
     <div class="footer" id="footer"></div>
-    <script src="sampler-settings.js"></script>
+    <script src="utils.js"></script>
+    <script src="settings.js"></script>
   </body>
 </html>
